requirements.txt
openai>=1.0.0
python-dotenv>=0.19.0
requests>=2.26.0
pydantic>=2.0.0
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
aiohttp>=3.8.0
websockets>=10.0
prisma>=0.9.0
cryptography>=3.4.7
pytest>=6.2.5
black>=22.3.0
isort>=5.9.3
flake8>=3.9.2
mypy>=0.910 